+++ air-script/tests/system/test_air.rs
@@ -3,8 +3,9 @@ use winter_math::fields::f64::BaseElement as Felt;
 use winterfell::{Trace, TraceTable};
 
 use crate::{
+    generate_air_test,
     helpers::{AirTester, MyTraceTable},
-    system::system::{PublicInputs, SystemAir},
+    system::system::PublicInputs,
 };
 
 #[derive(Clone)]
@@ -38,17 +39,4 @@ impl AirTester for SystemAirTester {
     }
 }
 
-#[test]
-fn test_system_air() {
-    let air_tester = Box::new(SystemAirTester {});
-    let length = 1024;
-
-    let main_trace = air_tester.build_main_trace(length);
-    let aux_trace = air_tester.build_aux_trace(length);
-    let pub_inputs = air_tester.public_inputs();
-    let trace_info = air_tester.build_trace_info(length);
-    let options = air_tester.build_proof_options();
-
-    let air = SystemAir::new(trace_info, pub_inputs, options);
-    main_trace.validate::<SystemAir, Felt>(&air, aux_trace.as_ref());
-}
+generate_air_test!(test_system_air, crate::system::system::SystemAir, SystemAirTester, 1024);